Updates
* Added "inrole" and "roleinfo" commands.

* Refactored ExtendedClientUtil.js.

* Overriden resolveUsers, resolveMembers, resolveChannels, resolveRoles and resolveGuilds of ClientUtil in ExtendedClientUtil. They will now support a 5th parameter called "tryExact" (boolean). When it's set to true, it will attempt to re-filter the filtered results to try to find an exact match.

* assert functions of ExtendeClientUtil are now Promise functions (assertUser, assertMember, assertRole, assertGuild and assertMemberOrUser).
assertChannel will come at a later date.

* Added multiSendEmbed function to ExtendedClientUtil.
It can be used to split embed with long description into multiple embeds. It can also wrap each description in code blocks, if necessary. It also has support for prefix and suffix (prefix will be prepended to the first embed and suffix will be appended to the last embed). This one is still very experimental though.
"inrole" command has been tested to list members of a role which had around 1000 members using the function. It could split it into 9 embeds just fine.

* multiSend function was improved to support firstMessage parameter. When it is set to a particular message, for the first message it will attempt to edit the said message instead of sending a new message.
On a side note, this parameter is also supported in multiSendEmbed function.

* Refactored userid.js, userinfo.js, ban.js, kick.js, avatar.js and lastfm.js.

* Added "--list" flag to "config" command. It can be used to list all valid keys for the configuration file. It will also list their respective settings if available (such as whether they keys are protected, what their default values are, what their allowed values are, etcetera).

* Added LightbringerError.js. It will extend Error by adding a timeout property. It will then be used by commandError.js to decided the delete timeout for the said error.

* Upgraded dependecies (the yarn.lock file).