Updates
* Added "guildinfo", "members" and "roleperms" commands.
guildinfo: shows informations of a currently viewed or a specific guild.
members: lists members of the currently viewed or a specific guild.
roleperms: lists permissions owned by a specific role (optionally from a specific guild).

* Updated "roleinfo" command to only list a maximum of 100 users by default. This can later be configured with "maxUsersListing" configuration key (using "config" command). This option will also affect "members" command.

* Updated some phrases in "userinfo" command. It will also now use 2048p link for the avatar link.

* Some more refactors to resolve and assert functions in ExtendedClientUtil.
Resolve for users, channels and guilds now support a 6th parameter called tryGlobalId (boolean). When it's set to true, it will try to get a user/channel/guild which have the exact same ID globally (not limited to a specific guild).
It's currently being used by assertUser, assertChannel and assertGuild functions.

* console.debug() will now set the inspection depth to 0 instead of 2.