Updates
* Upgraded dependencies (yarn.lock) as well as their version requirements in package.json file.

* Added "booru" (using BobbyWibowo/booru -- a modified version of the original to use Snekfetch instead of Request) and "encodeurl" to dependencies.

* Added "booru" command.
This command can be used to fetch random images from various booru sites (it will fetch from gelbooru by default).
The aformentioned dependencies were added for this command.
And of course, you can specify tags for the search keywords.
On a side note, I've only tried this command with gelbooru (using up to 3 tags) and lolibooru (using up to 2 tags).

* Added a note to "setup.js" file stating that it's currently not needed to initialize the bot using the file.

* Removed a stray ")" character in "emojis" command.

* Refactored "help" command. Hopefully the code will appear more readable now.

* Updated ready.js to use stripIndent instead of stripIndents (check out the npm page of common-tags to see more information of the both of them).

* Added cleanUrl() function to ExtendedClientUtil.js. This function can be used to encode then clean URL (replacing whitespace with "+", etcetera).