Updates
* Updated "booru" commands.
Added "--liteMode" flag to toggle lite mode. When lite mode is turned on, the command will no longer print extra information such as scores, ratings and sources.
"--defaultSite" option now have an extra alias, "-d=".
And some other attention given to the logics.

* Added a new command: "boorubomb".
This command allows you to show a total of 4 random images from different booru sites. You can customize which booru sites to use. In fact, it's not limited to 4 sites, but even more! Unlike "booru" command, this will only print the image URLs.

* Updated some commands to allow quoted value for options (something like: --guild="guild name here").

* Renamed getGuildColors util function to getGuildColor (merely removed the "s" at the end).

* Added a new util function called getAverageColor to get the average color of an image. It only accepts Buffer, Path or URL of PNG, JPG or BMP images. For now it's being used by getGuildColor util function and a new command called "averagecolor" (info below).

* Added a new command: "averagecolor".
With this command, you can calculate the average color of an image. You only need to pass an URL of the image to the command.

* Added a new command: "discriminator".
With this command you can list users which have a certain discriminator (it will only lists users that are currently cached by the bot).

* Improved logics of "upload" command.

* And maybe some others ...