Updated eval.js
Run "eval()" outside of the class structure so that it's not in strict mode. It will still have access to the class through "this", so it should be similar to its old behavior, other than the fact that it's running in global scope (which allows you to do variable assignments in global scope; should be useful for testing purposes).